requirements.txt
torch>=1.12.0
torchvision
numpy
Pillow
opencv-python
git+https://github.com/facebookresearch/segment-anything.git
svgwrite
